Digital_clock
This project, titled "Digital Clock," is a web application designed to display the current time in a digital format. It utilizes HTML for structuring the content, CSS for styling the clock's appearance, and JavaScript for dynamically updating the time in real-time. The clock's design and functionality can be customized to suit various preferences and applications.